+++ R/cor_test.R
@@ -4,17 +4,51 @@
 #'
 #' @param data A data frame.
 #' @param x,y Names of two variables present in the data.
-#' @param ci Confidence/Credible Interval level. If \code{"default"}, then it is set to 0.95 (95\% CI).
-#' @param method A character string indicating which correlation coefficient is to be used for the test. One of \code{"pearson"} (default), \code{"kendall"}, \code{"spearman"}, \code{"biserial"}, \code{"polychoric"}, \code{"tetrachoric"}, \code{"biweight"}, \code{"distance"}, \code{"percentage"} (for percentage bend correlation), \code{"blomqvist"} (for Blomqvist's coefficient), \code{"hoeffding"} (for Hoeffding's D), \code{"gamma"}, \code{"gaussian"} (for Gaussian Rank correlation) or \code{"shepherd"} (for Shepherd's Pi correlation). Setting \code{"auto"} will attempt at selecting the most relevant method (polychoric when ordinal factors involved, tetrachoric when dichotomous factors involved, point-biserial if one dichotomous and one continuous and pearson otherwise).
-#' @param bayesian,partial_bayesian If TRUE, will run the correlations under a Bayesian framework. Note that for partial correlations, you will also need to set \code{partial_bayesian} to \code{TRUE} to obtain "full" Bayesian partial correlations. Otherwise, you will obtain pseudo-Bayesian partial correlations (i.e., Bayesian correlation based on frequentist partialization).
-#' @param include_factors If \code{TRUE}, the factors are kept and eventually converted to numeric or used as random effects (depending of \code{multilevel}). If \code{FALSE}, factors are removed upfront.
-#' @param partial Can be \code{TRUE} or \code{"semi"} for partial and semi-partial correlations, respectively.
+#' @param ci Confidence/Credible Interval level. If \code{"default"}, then it is
+#'   set to 0.95 (95\% CI).
+#' @param method A character string indicating which correlation coefficient is
+#'   to be used for the test. One of \code{"pearson"} (default),
+#'   \code{"kendall"}, \code{"spearman"}, \code{"biserial"},
+#'   \code{"polychoric"}, \code{"tetrachoric"}, \code{"biweight"},
+#'   \code{"distance"}, \code{"percentage"} (for percentage bend correlation),
+#'   \code{"blomqvist"} (for Blomqvist's coefficient), \code{"hoeffding"} (for
+#'   Hoeffding's D), \code{"gamma"}, \code{"gaussian"} (for Gaussian Rank
+#'   correlation) or \code{"shepherd"} (for Shepherd's Pi correlation). Setting
+#'   \code{"auto"} will attempt at selecting the most relevant method
+#'   (polychoric when ordinal factors involved, tetrachoric when dichotomous
+#'   factors involved, point-biserial if one dichotomous and one continuous and
+#'   pearson otherwise).
+#' @param bayesian,partial_bayesian If TRUE, will run the correlations under a
+#'   Bayesian framework. Note that for partial correlations, you will also need
+#'   to set \code{partial_bayesian} to \code{TRUE} to obtain "full" Bayesian
+#'   partial correlations. Otherwise, you will obtain pseudo-Bayesian partial
+#'   correlations (i.e., Bayesian correlation based on frequentist
+#'   partialization).
+#' @param include_factors If \code{TRUE}, the factors are kept and eventually
+#'   converted to numeric or used as random effects (depending of
+#'   \code{multilevel}). If \code{FALSE}, factors are removed upfront.
+#' @param partial Can be \code{TRUE} or \code{"semi"} for partial and
+#'   semi-partial correlations, respectively.
 #' @inheritParams effectsize::adjust
-#' @param bayesian_prior For the prior argument, several named values are recognized: \code{"medium.narrow"}, \code{"medium"}, \code{"wide"}, and \code{"ultrawide"}. These correspond to scale values of \code{1/sqrt(27)}, \code{1/3}, \code{1/sqrt(3)} and \code{1}, respectively. See the \code{BayesFactor::correlationBF} function.
-#' @param bayesian_ci_method,bayesian_test See arguments in \code{\link[=parameters]{model_parameters}} for \code{BayesFactor} tests.
-#' @param robust If TRUE, will rank-transform the variables prior to estimating the correlation. Note that, for instance, a Pearson's correlation on rank-transformed data is equivalent to a Spearman's rank correlation. Thus, using \code{robust=TRUE} and \code{method="spearman"} is redundant. Nonetheless, it is an easy way to increase the robustness of the correlation (as well as obtaining Bayesian Spearman rank Correlations).
-#' @param winsorize Either \code{FALSE} or a number between 0 and 1 (e.g., \code{0.2}) that corresponds to the threshold of desired \code{\link[=winsorize]{winsorization}}.
-#' @param ... Arguments passed to or from other methods.
+#' @param bayesian_prior For the prior argument, several named values are
+#'   recognized: \code{"medium.narrow"}, \code{"medium"}, \code{"wide"}, and
+#'   \code{"ultrawide"}. These correspond to scale values of \code{1/sqrt(27)},
+#'   \code{1/3}, \code{1/sqrt(3)} and \code{1}, respectively. See the
+#'   \code{BayesFactor::correlationBF} function.
+#' @param bayesian_ci_method,bayesian_test See arguments in
+#'   \code{\link[=parameters]{model_parameters}} for \code{BayesFactor} tests.
+#' @param robust If \code{TRUE}, will rank-transform the variables prior to
+#'   estimating the correlation. Note that, for instance, a Pearson's
+#'   correlation on rank-transformed data is equivalent to a Spearman's rank
+#'   correlation. Thus, using \code{robust=TRUE} and \code{method="spearman"} is
+#'   redundant. Nonetheless, it is an easy way to increase the robustness of the
+#'   correlation (as well as obtaining Bayesian Spearman rank Correlations).
+#' @param winsorize Either \code{FALSE} or a number between 0 and 1 (e.g.,
+#'   \code{0.2}) that corresponds to the threshold of desired
+#'   \code{\link[=winsorize]{winsorization}}.
+#' @param verbose Toggle warnings.
+#' @param ... Additional arguments (e.g., \code{alternative}) to be passed to
+#'   other methods. See \code{stats::cor.test} for further details.
 #'
 #'
 #' @inherit correlation details
@@ -30,29 +64,37 @@
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "biweight")
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "distance")
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "percentage")
-#' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "blomqvist")
-#' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "hoeffding")
+#' if (require("wdm", quietly = TRUE)) {
+#'   cor_test(iris, "Sepal.Length", "Sepal.Width", method = "blomqvist")
+#' }
+#' if (require("Hmisc", quietly = TRUE)) {
+#'   cor_test(iris, "Sepal.Length", "Sepal.Width", method = "hoeffding")
+#' }
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "gamma")
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "gaussian")
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "shepherd")
-#' cor_test(iris, "Sepal.Length", "Sepal.Width", bayesian = TRUE)
+#' if (require("BayesFactor", quietly = TRUE)) {
+#'   cor_test(iris, "Sepal.Length", "Sepal.Width", bayesian = TRUE)
+#' }
 #'
 #' # Tetrachoric
-#' data <- iris
-#' data$Sepal.Width_binary <- ifelse(data$Sepal.Width > 3, 1, 0)
-#' data$Petal.Width_binary <- ifelse(data$Petal.Width > 1.2, 1, 0)
-#' cor_test(data, "Sepal.Width_binary", "Petal.Width_binary", method = "tetrachoric")
+#' if (require("psych", quietly = TRUE)) {
+#'   data <- iris
+#'   data$Sepal.Width_binary <- ifelse(data$Sepal.Width > 3, 1, 0)
+#'   data$Petal.Width_binary <- ifelse(data$Petal.Width > 1.2, 1, 0)
+#'   cor_test(data, "Sepal.Width_binary", "Petal.Width_binary", method = "tetrachoric")
 #'
-#' # Biserial
-#' cor_test(data, "Sepal.Width", "Petal.Width_binary", method = "biserial")
+#'   # Biserial
+#'   cor_test(data, "Sepal.Width", "Petal.Width_binary", method = "biserial")
 #'
-#' # Polychoric
-#' data$Petal.Width_ordinal <- as.factor(round(data$Petal.Width))
-#' data$Sepal.Length_ordinal <- as.factor(round(data$Sepal.Length))
-#' cor_test(data, "Petal.Width_ordinal", "Sepal.Length_ordinal", method = "polychoric")
+#'   # Polychoric
+#'   data$Petal.Width_ordinal <- as.factor(round(data$Petal.Width))
+#'   data$Sepal.Length_ordinal <- as.factor(round(data$Sepal.Length))
+#'   cor_test(data, "Petal.Width_ordinal", "Sepal.Length_ordinal", method = "polychoric")
 #'
-#' # When one variable is continuous, will run 'polyserial' correlation
-#' cor_test(data, "Sepal.Width", "Sepal.Length_ordinal", method = "polychoric")
+#'   # When one variable is continuous, will run 'polyserial' correlation
+#'   cor_test(data, "Sepal.Width", "Sepal.Length_ordinal", method = "polychoric")
+#' }
 #'
 #' # Robust (these two are equivalent)
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", method = "pearson", robust = TRUE)
@@ -60,17 +102,15 @@
 #'
 #' # Winsorized
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", winsorize = 0.2)
-#'
+#' \dontrun{
 #' # Partial
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", partial = TRUE)
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", multilevel = TRUE)
-#' \donttest{
 #' cor_test(iris, "Sepal.Length", "Sepal.Width", partial_bayesian = TRUE)
 #' }
-#'
 #' @importFrom effectsize adjust ranktransform
 #' @importFrom parameters data_to_numeric
-#' @importFrom stats complete.cases
+#' @importFrom stats complete.cases na.omit sd
 #' @export
 cor_test <- function(data,
                      x,
@@ -87,10 +127,14 @@ cor_test <- function(data,
                      multilevel = FALSE,
                      robust = FALSE,
                      winsorize = FALSE,
+                     verbose = TRUE,
                      ...) {
 
   # Sanity checks
-  if (!x %in% names(data) | !y %in% names(data)) stop("The names you entered for x and y are not available in the dataset. Make sure there are no typos!")
+  if (!x %in% names(data) | !y %in% names(data)) {
+    stop("The names you entered for x and y are not available in the dataset. Make sure there are no typos!")
+  }
+
   if (ci == "default") ci <- 0.95
   if (partial == FALSE & (partial_bayesian | multilevel)) partial <- TRUE
 
@@ -106,8 +150,16 @@ cor_test <- function(data,
   }
 
   # Winsorize
-  if (!isFALSE(winsorize)) {
-    data[c(x, y)] <- winsorize(data[c(x, y)], threshold = winsorize)
+  if (!isFALSE(winsorize) && !is.null(winsorize)) {
+    # set default
+    if (isTRUE(winsorize)) {
+      winsorize <- .1
+    }
+
+    # winsorization would otherwise fail in case of NAs present
+    data <- stats::na.omit(data)
+
+    data[c(x, y)] <- winsorize(data[c(x, y)], threshold = winsorize, verbose = verbose)
   }
 
   # Robust
@@ -119,7 +171,9 @@ cor_test <- function(data,
   # This is a trick in case the number of valid observations is lower than 3
   invalid <- FALSE
   if (n_obs < 3) {
-    warning(paste(x, "and", y, "have less than 3 complete observations. Returning NA."))
+    if (isTRUE(verbose)) {
+      warning(paste(x, "and", y, "have less than 3 complete observations. Returning NA."), call. = FALSE)
+    }
     invalid <- TRUE
     original_info <- list(data = data, x = x, y = y)
     data <- datasets::mtcars # Basically use a working dataset so the correlation doesn't fail
@@ -186,7 +240,17 @@ cor_test <- function(data,
     } else if (method %in% c("shepherd", "sheperd", "shepherdspi", "pi")) {
       out <- .cor_test_shepherd(data, x, y, ci = ci, bayesian = TRUE, ...)
     } else {
-      out <- .cor_test_bayes(data, x, y, ci = ci, method = method, bayesian_prior = bayesian_prior, bayesian_ci_method = bayesian_ci_method, bayesian_test = bayesian_test, ...)
+      out <- .cor_test_bayes(
+        data,
+        x,
+        y,
+        ci = ci,
+        method = method,
+        bayesian_prior = bayesian_prior,
+        bayesian_ci_method = bayesian_ci_method,
+        bayesian_test = bayesian_test,
+        ...
+      )
     }
   }
 
@@ -198,12 +262,13 @@ cor_test <- function(data,
     out[!names(out) %in% c("Parameter1", "Parameter2")] <- NA
   }
 
-  # Number of observations
+  # Number of observations and CI
   out$n_Obs <- n_obs
+  out$CI <- ci
 
   # Reorder columns
   if ("CI_low" %in% names(out)) {
-    order <- c("Parameter1", "Parameter2", "r", "rho", "Dxy", "CI_low", "CI_high")
+    order <- c("Parameter1", "Parameter2", "r", "rho", "tau", "Dxy", "CI", "CI_low", "CI_high")
     out <- out[c(order[order %in% names(out)], setdiff(colnames(out), order[order %in% names(out)]))]
   }